DiscordBot/requirements.txt
discord.py
python-dotenv
sqlalchemy
alembic
aiosqlite
psycopg2-binary
supabase-py